Add GetAllFiles method and RemoteFileInfo class
A new method `GetAllFiles` was added to the `IStorageProvider` interface in the `Octockup.Server.Providers` namespace. The `ScpProvider` class now includes an implementation of the `GetAllFiles` method, which currently throws a `NotImplementedException`. Additionally, a new class `RemoteFileInfo` was added to the `Octockup.Server.Providers` namespace, including properties for `Name`, `Path`, `Size`, and `LastModified` to represent information about remote files.