Simplify Dockerfile by combining installation commands
The Dockerfile has been modified to streamline the installation of
`nodejs`, `curl`, and `git-lfs`. Previously, the installation commands
for `curl` and `git-lfs` were separate from the `nodejs` installation.
The updated Dockerfile combines these installations into a single `RUN`
command, which simplifies the Dockerfile and potentially reduces the
number of layers in the resulting Docker image.