Add Elapsed property to BackupSnapshot and update schema
Introduce a new Elapsed property of type TimeSpan to the BackupSnapshot class in the Octockup.Server.Database namespace, mapped to the elapsed column in the database. Update the BackupRequestHandler class in the Octockup.Server.Handlers namespace to set the Elapsed property of the snapshot object from the progressTracker at two points. Modify the PostgresDbContextModelSnapshot class in the Octockup.Server.Migrations.PostgresMigrations namespace to include the new Elapsed property for the BackupSnapshot entity, mapping it to the elapsed column with the interval type. Add a new migration class 20250110054142_AddElapsed to the Octockup.Server.Migrations.PostgresMigrations namespace, adding the elapsed column to the backup_snapshots table with a default value of TimeSpan.Zero. Include the corresponding designer file 20250110054142_AddElapsed.Designer.cs with auto-generated code to build the target model for the migration, reflecting the addition of the elapsed column to the backup_snapshots table.