Handle S3 checksum errors and improve DeleteAsync formatting
Updated `GetObjectAsync` to include `ChecksumMode` set to
"DISABLED" in the `GetObjectRequest`. Added a `try-catch` block
to handle `ArgumentOutOfRangeException` for checksum errors.
If `_validateChecksums` is false, a pre-signed URL is generated
and used to fetch the object. Re-throws the exception if
conditions are not met.

Minor formatting improvement in `DeleteAsync` by removing an
unnecessary blank line.